Replace old delete command with new delete commands
