+++ src/main/java/seedu/hirehub/logic/LogicManager.java
@@ -12,11 +12,7 @@ import seedu.hirehub.logic.commands.Command;
 import seedu.hirehub.logic.commands.CommandResult;
 import seedu.hirehub.logic.commands.exceptions.CommandException;
 import seedu.hirehub.logic.parser.AddressBookParser;
-import seedu.hirehub.logic.parser.ClearConfirmationStageParser;
 import seedu.hirehub.logic.parser.ConfirmationStageParser;
-import seedu.hirehub.logic.parser.DeleteApplicationConfirmationStageParser;
-import seedu.hirehub.logic.parser.DeleteConfirmationStageParser;
-import seedu.hirehub.logic.parser.DeleteJobConfirmationStageParser;
 import seedu.hirehub.logic.parser.exceptions.ParseException;
 import seedu.hirehub.model.Model;
 import seedu.hirehub.model.ReadOnlyAddressBook;
@@ -32,7 +28,7 @@ public class LogicManager implements Logic {
     public static final String FILE_OPS_ERROR_FORMAT = "Could not save data due to the following error: %s";
 
     public static final String FILE_OPS_PERMISSION_ERROR_FORMAT =
-            "Could not save data to file %s due to insufficient permissions to write to the file or the folder.";
+        "Could not save data to file %s due to insufficient permissions to write to the file or the folder.";
 
     private final Logger logger = LogsCenter.getLogger(LogicManager.class);
 
@@ -40,10 +36,6 @@ public class LogicManager implements Logic {
     private final Storage storage;
     private final AddressBookParser addressBookParser;
     private final ConfirmationStageParser confirmationStageParser;
-    private final ClearConfirmationStageParser clearConfirmationStageParser;
-    private final DeleteConfirmationStageParser deleteConfirmationStageParser;
-    private final DeleteApplicationConfirmationStageParser deleteApplicationConfirmationStageParser;
-    private final DeleteJobConfirmationStageParser deleteJobConfirmationStageParser;
     private CommandBoxState state;
 
     /**
@@ -54,10 +46,6 @@ public class LogicManager implements Logic {
         this.storage = storage;
         addressBookParser = new AddressBookParser();
         confirmationStageParser = new ConfirmationStageParser();
-        clearConfirmationStageParser = new ClearConfirmationStageParser();
-        deleteConfirmationStageParser = new DeleteConfirmationStageParser();
-        deleteApplicationConfirmationStageParser = new DeleteApplicationConfirmationStageParser();
-        deleteJobConfirmationStageParser = new DeleteJobConfirmationStageParser();
         this.state = CommandBoxState.NORMAL;
     }
 
@@ -73,18 +61,6 @@ public class LogicManager implements Logic {
             break;
         case CONFIRM:
             command = confirmationStageParser.parseCommand(commandText);
-            break;
-        case CLEARCONFIRM:
-            command = clearConfirmationStageParser.parseCommand(commandText);
-            break;
-        case DELETECONFIRM:
-            command = deleteConfirmationStageParser.parseCommand(commandText);
-            break;
-        case DELETEAPPLICATIONCONFIRM:
-            command = deleteApplicationConfirmationStageParser.parseCommand(commandText);
-            break;
-        case DELETEJOBCONFIRM:
-            command = deleteJobConfirmationStageParser.parseCommand(commandText);
             break;
         default:
             break;
