--- src/main/java/seedu/hirehub/logic/commands/AbortClearCommand.java
@@ -1,19 +1,0 @@
-package seedu.hirehub.logic.commands;
-
-import static java.util.Objects.requireNonNull;
-
-import seedu.hirehub.logic.CommandBoxState;
-import seedu.hirehub.model.Model;
-
-/**
- * Aborts the clearing of the database.
- */
-public class AbortClearCommand extends Command {
-    public static final String MESSAGE_SUCCESS = "Clearing aborted";
-
-    @Override
-    public CommandResult execute(Model model) {
-        requireNonNull(model);
-        return new CommandResult(MESSAGE_SUCCESS, CommandBoxState.NORMAL);
-    }
-}
