+++ src/main/java/seedu/hirehub/logic/commands/ClearCommand.java
@@ -7,19 +7,37 @@ import seedu.hirehub.model.AddressBook;
 import seedu.hirehub.model.Model;
 
 /**
- * Clears the address book.
+ * Pushes the program into the state to clear the database.
  */
-public class ClearCommand extends Command {
-
+public class ClearCommand extends AbstractDeleteCommand {
     public static final String COMMAND_WORD = "clear";
-    public static final String MESSAGE_SUCCESS = "Address book has been cleared!";
+    public static final String MESSAGE_CONFIRM_STAGE = "Clearing the database is irreversible. Proceed? (Y/N)";
+    public static final String MESSAGE_DATABASE_EMPTY = "The database is empty.";
 
+    public static final String MESSAGE_CLEAR_CONFIRM = "Address book has been cleared!";
+
+    public static final String MESSAGE_CLEAR_CANCEL = "Clearing aborted";
 
     @Override
     public CommandResult execute(Model model) {
         requireNonNull(model);
+        if (model.getAddressBook().getPersonList().isEmpty()) {
+            return new CommandResult(MESSAGE_DATABASE_EMPTY, CommandBoxState.NORMAL);
+        }
+        return new CommandResult(MESSAGE_CONFIRM_STAGE, CommandBoxState.CONFIRM);
+    }
+
+    @Override
+    public String performDelete(Model model) {
+        requireNonNull(model);
         model.setAddressBook(new AddressBook());
         model.clearApplications();
-        return new CommandResult(MESSAGE_SUCCESS, CommandBoxState.NORMAL);
+        return MESSAGE_CLEAR_CONFIRM;
+
+    }
+
+    @Override
+    public String denyDelete(Model model) {
+        return MESSAGE_CLEAR_CANCEL;
     }
 }
