--- src/main/java/seedu/hirehub/logic/commands/DeleteCommand.java
@@ -1,52 +1,0 @@
-package seedu.hirehub.logic.commands;
-
-import static java.util.Objects.requireNonNull;
-
-import seedu.hirehub.commons.util.ToStringBuilder;
-import seedu.hirehub.logic.Messages;
-import seedu.hirehub.logic.commands.exceptions.CommandException;
-import seedu.hirehub.model.Model;
-import seedu.hirehub.model.person.Person;
-
-/**
- * Deletes a person identified using it's displayed index from the address book.
- */
-public class DeleteCommand extends Command {
-
-    public static final String COMMAND_WORD = "delete";
-
-    public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Deletes the person identified by the index number used in the displayed person list.\n"
-            + "Parameters: INDEX (must be a positive integer)\n"
-            + "Example: " + COMMAND_WORD + " 1";
-
-    public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
-
-    @Override
-    public CommandResult execute(Model model) throws CommandException {
-        requireNonNull(model);
-        Person personToDelete = model.getLastMentionedPerson().get();
-        model.deletePerson(personToDelete);
-        model.removeApplications(personToDelete);
-        return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof DeleteCommand)) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .toString();
-    }
-}
