+++ src/main/java/seedu/hirehub/logic/parser/AddressBookParser.java
@@ -11,8 +11,12 @@ import seedu.hirehub.commons.core.LogsCenter;
 import seedu.hirehub.logic.commands.AddApplicationCommand;
 import seedu.hirehub.logic.commands.AddCommand;
 import seedu.hirehub.logic.commands.AddJobCommand;
+import seedu.hirehub.logic.commands.ClearCommand;
 import seedu.hirehub.logic.commands.Command;
 import seedu.hirehub.logic.commands.CommentCommand;
+import seedu.hirehub.logic.commands.DeleteApplicationCommand;
+import seedu.hirehub.logic.commands.DeleteJobCommand;
+import seedu.hirehub.logic.commands.DeletePersonCommand;
 import seedu.hirehub.logic.commands.DeleteTagCommand;
 import seedu.hirehub.logic.commands.EditCommand;
 import seedu.hirehub.logic.commands.EditJobCommand;
@@ -22,10 +26,6 @@ import seedu.hirehub.logic.commands.HelpCommand;
 import seedu.hirehub.logic.commands.ListApplicationCommand;
 import seedu.hirehub.logic.commands.ListCommand;
 import seedu.hirehub.logic.commands.ListJobCommand;
-import seedu.hirehub.logic.commands.NewClearCommand;
-import seedu.hirehub.logic.commands.NewDeleteApplicationCommand;
-import seedu.hirehub.logic.commands.NewDeleteJobCommand;
-import seedu.hirehub.logic.commands.NewDeletePersonCommand;
 import seedu.hirehub.logic.commands.SearchApplicationCommand;
 import seedu.hirehub.logic.commands.SearchCommand;
 import seedu.hirehub.logic.commands.SearchJobCommand;
@@ -74,11 +74,11 @@ public class AddressBookParser {
         case EditCommand.COMMAND_WORD:
             return new EditCommandParser().parse(arguments);
 
-        case NewDeletePersonCommand.COMMAND_WORD:
-            return new InitDeleteCommandParser().parse(arguments);
+        case DeletePersonCommand.COMMAND_WORD:
+            return new DeletePersonCommandParser().parse(arguments);
 
-        case NewClearCommand.COMMAND_WORD:
-            return new NewClearCommand();
+        case ClearCommand.COMMAND_WORD:
+            return new ClearCommand();
 
         case SearchCommand.COMMAND_WORD:
             return new SearchCommandParser().parse(arguments);
@@ -113,8 +113,8 @@ public class AddressBookParser {
         case EditJobCommand.COMMAND_WORD:
             return new EditJobCommandParser().parse(arguments);
 
-        case NewDeleteJobCommand.COMMAND_WORD:
-            return new InitDeleteJobCommandParser().parse(arguments);
+        case DeleteJobCommand.COMMAND_WORD:
+            return new DeleteJobCommandParser().parse(arguments);
 
         case SearchJobCommand.COMMAND_WORD:
             return new SearchJobCommandParser().parse(arguments);
@@ -122,8 +122,8 @@ public class AddressBookParser {
         case AddApplicationCommand.COMMAND_WORD:
             return new AddApplicationCommandParser().parse(arguments);
 
-        case NewDeleteApplicationCommand.COMMAND_WORD:
-            return new InitDeleteApplicationCommandParser().parse(arguments);
+        case DeleteApplicationCommand.COMMAND_WORD:
+            return new DeleteApplicationCommandParser().parse(arguments);
 
         case SearchApplicationCommand.COMMAND_WORD:
             return new SearchApplicationCommandParser().parse(arguments);
