+++ src/main/java/seedu/hirehub/logic/parser/DeleteApplicationCommandParser.java
@@ -4,24 +4,23 @@ import static seedu.hirehub.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.hirehub.commons.core.index.Index;
 import seedu.hirehub.logic.commands.DeleteApplicationCommand;
-import seedu.hirehub.logic.commands.NewDeleteApplicationCommand;
 import seedu.hirehub.logic.parser.exceptions.ParseException;
 
 /**
- * Parses input arguments and creates a new DeleteCommand object
+ * Parses input arguments and creates a new DeleteApplicationCommand object
  */
-public class InitDeleteApplicationCommandParser implements Parser<NewDeleteApplicationCommand> {
+public class DeleteApplicationCommandParser implements Parser<DeleteApplicationCommand> {
 
     /**
-     * Parses the given {@code String} of arguments in the context of the DeleteCommand
-     * and returns a DeleteCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the DeleteApplicationCommand
+     * and returns a DeleteApplicationCommand object for execution.
      *
      * @throws ParseException if the user input does not conform the expected format
      */
-    public NewDeleteApplicationCommand parse(String args) throws ParseException {
+    public DeleteApplicationCommand parse(String args) throws ParseException {
         try {
             Index index = ParserUtil.parseIndex(args);
-            return new NewDeleteApplicationCommand(index);
+            return new DeleteApplicationCommand(index);
         } catch (ParseException pe) {
             throw new ParseException(
                 String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteApplicationCommand.MESSAGE_USAGE), pe);
