+++ src/main/java/seedu/hirehub/logic/parser/DeleteJobCommandParser.java
@@ -3,27 +3,27 @@ package seedu.hirehub.logic.parser;
 import static seedu.hirehub.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.hirehub.commons.core.index.Index;
-import seedu.hirehub.logic.commands.NewDeleteJobCommand;
+import seedu.hirehub.logic.commands.DeleteJobCommand;
 import seedu.hirehub.logic.parser.exceptions.ParseException;
 
 /**
  * Parses input arguments and creates a new DeleteJobCommand object
  */
-public class InitDeleteJobCommandParser implements Parser<NewDeleteJobCommand> {
+public class DeleteJobCommandParser implements Parser<DeleteJobCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the DeleteJobCommand
-     * and returns a DeleteCommand object for execution.
+     * and returns a DeleteJobCommand object for execution.
      *
      * @throws ParseException if the user input does not conform the expected format
      */
-    public NewDeleteJobCommand parse(String args) throws ParseException {
+    public DeleteJobCommand parse(String args) throws ParseException {
         try {
             Index index = ParserUtil.parseIndex(args);
-            return new NewDeleteJobCommand(index);
+            return new DeleteJobCommand(index);
         } catch (ParseException pe) {
             throw new ParseException(
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, NewDeleteJobCommand.MESSAGE_USAGE), pe);
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteJobCommand.MESSAGE_USAGE), pe);
         }
     }
