--- src/main/java/seedu/hirehub/logic/parser/DeleteJobConfirmationStageParser.java
@@ -1,58 +1,0 @@
-package seedu.hirehub.logic.parser;
-
-import static seedu.hirehub.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.hirehub.logic.Messages.MESSAGE_UNKNOWN_COMMAND_DELETE_CONFIRMATION;
-
-import java.util.logging.Logger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import seedu.hirehub.commons.core.LogsCenter;
-import seedu.hirehub.logic.commands.AbortDeleteJobCommand;
-import seedu.hirehub.logic.commands.Command;
-import seedu.hirehub.logic.commands.DeleteJobCommand;
-import seedu.hirehub.logic.commands.HelpCommand;
-import seedu.hirehub.logic.parser.exceptions.ParseException;
-
-/**
- * Parses the user input when the user intends to delete a job.
- */
-public class DeleteJobConfirmationStageParser {
-    /**
-     * Used for initial separation of command word and args.
-     */
-    private static final Pattern BASIC_COMMAND_FORMAT = Pattern.compile("(?<commandWord>\\S+)(?<arguments>.*)");
-    private static final Logger logger = LogsCenter.getLogger(ClearConfirmationStageParser.class);
-
-    /**
-     * Parses user input into command for execution.
-     *
-     * @param userInput full user input string
-     * @return the command based on the user input
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    public Command parseCommand(String userInput) throws ParseException {
-        final Matcher matcher = BASIC_COMMAND_FORMAT.matcher(userInput.trim());
-        if (!matcher.matches()) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, HelpCommand.MESSAGE_USAGE));
-        }
-
-        final String commandWord = matcher.group("commandWord");
-        final String arguments = matcher.group("arguments");
-
-        // Note to developers: Change the log level in config.json to enable lower level (i.e., FINE, FINER and lower)
-        // log messages such as the one below.
-        // Lower level log messages are used sparingly to minimize noise in the code.
-        logger.fine("Command word: " + commandWord + "; Arguments: " + arguments);
-
-        switch (commandWord) {
-        case "Y":
-            return new DeleteJobCommand();
-        case "N":
-            return new AbortDeleteJobCommand();
-        default:
-            logger.finer("This user input caused a ParseException: " + userInput);
-            throw new ParseException(MESSAGE_UNKNOWN_COMMAND_DELETE_CONFIRMATION);
-        }
-    }
-}
