+++ src/main/java/seedu/hirehub/logic/parser/DeletePersonCommandParser.java
@@ -1,0 +1,30 @@
+package seedu.hirehub.logic.parser;
+
+import static seedu.hirehub.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import seedu.hirehub.commons.core.index.Index;
+import seedu.hirehub.logic.commands.DeletePersonCommand;
+import seedu.hirehub.logic.parser.exceptions.ParseException;
+
+/**
+ * Parses input arguments and creates a new DeletePersonCommand object
+ */
+public class DeletePersonCommandParser implements Parser<DeletePersonCommand> {
+
+    /**
+     * Parses the given {@code String} of arguments in the context of the DeletePersonCommand
+     * and returns a DeletePersonCommand object for execution.
+     *
+     * @throws ParseException if the user input does not conform the expected format
+     */
+    public DeletePersonCommand parse(String args) throws ParseException {
+        try {
+            Index index = ParserUtil.parseIndex(args);
+            return new DeletePersonCommand(index);
+        } catch (ParseException pe) {
+            throw new ParseException(
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeletePersonCommand.MESSAGE_USAGE), pe);
+        }
+    }
+
+}
