--- src/main/java/seedu/hirehub/logic/parser/InitDeleteCommandParser.java
@@ -1,30 +1,0 @@
-package seedu.hirehub.logic.parser;
-
-import static seedu.hirehub.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-
-import seedu.hirehub.commons.core.index.Index;
-import seedu.hirehub.logic.commands.NewDeletePersonCommand;
-import seedu.hirehub.logic.parser.exceptions.ParseException;
-
-/**
- * Parses input arguments and creates a new DeleteCommand object
- */
-public class InitDeleteCommandParser implements Parser<NewDeletePersonCommand> {
-
-    /**
-     * Parses the given {@code String} of arguments in the context of the DeleteCommand
-     * and returns a DeleteCommand object for execution.
-     *
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    public NewDeletePersonCommand parse(String args) throws ParseException {
-        try {
-            Index index = ParserUtil.parseIndex(args);
-            return new NewDeletePersonCommand(index);
-        } catch (ParseException pe) {
-            throw new ParseException(
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, NewDeletePersonCommand.MESSAGE_USAGE), pe);
-        }
-    }
-
-}
