+++ src/test/java/seedu/hirehub/logic/commands/ClearCommandTest.java
@@ -18,7 +18,8 @@ public class ClearCommandTest {
         Model model = new ModelManager();
         Model expectedModel = new ModelManager();
 
-        assertCommandSuccess(new ClearCommand(), model, ClearCommand.MESSAGE_SUCCESS, expectedModel);
+        assertCommandSuccess(new ClearCommand().whenConfirmed()
+            , model, ClearCommand.MESSAGE_CLEAR_CONFIRM, expectedModel);
     }
 
     @Test
@@ -27,7 +28,8 @@ public class ClearCommandTest {
         Model expectedModel = new ModelManager(getTypicalAddressBook(), new UniqueJobList(), new UserPrefs());
         expectedModel.setAddressBook(new AddressBook());
 
-        assertCommandSuccess(new ClearCommand(), model, ClearCommand.MESSAGE_SUCCESS, expectedModel);
+        assertCommandSuccess(new ClearCommand().whenConfirmed()
+            , model, ClearCommand.MESSAGE_CLEAR_CONFIRM, expectedModel);
     }
 
 }
