--- src/test/java/seedu/hirehub/logic/commands/InitDeleteCommandTest.java
@@ -1,101 +1,0 @@
-package seedu.hirehub.logic.commands;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.hirehub.logic.commands.CommandTestUtil.assertCommandFailure;
-import static seedu.hirehub.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.hirehub.logic.commands.CommandTestUtil.showPersonAtIndex;
-import static seedu.hirehub.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
-import static seedu.hirehub.testutil.TypicalIndexes.INDEX_SECOND_PERSON;
-import static seedu.hirehub.testutil.TypicalPersons.getTypicalAddressBook;
-
-import org.junit.jupiter.api.Test;
-
-import seedu.hirehub.commons.core.index.Index;
-import seedu.hirehub.logic.Messages;
-import seedu.hirehub.model.Model;
-import seedu.hirehub.model.ModelManager;
-import seedu.hirehub.model.UserPrefs;
-import seedu.hirehub.model.job.UniqueJobList;
-
-/**
- * Contains integration tests (interaction with the Model) and unit tests for
- * {@code DeleteCommand}.
- */
-public class InitDeleteCommandTest {
-
-    private Model model = new ModelManager(getTypicalAddressBook(), new UniqueJobList(), new UserPrefs());
-
-    @Test
-    public void execute_validIndexUnfilteredList_success() {
-        InitDeleteCommand deleteCommand = new InitDeleteCommand(INDEX_FIRST_PERSON);
-
-        String expectedMessage = String.format(InitDeleteCommand.MESSAGE_SUCCESS);
-
-        ModelManager expectedModel = new ModelManager(model.getAddressBook(),
-            new UniqueJobList(), new UserPrefs());
-
-        assertCommandSuccess(deleteCommand, model, expectedMessage, expectedModel);
-    }
-
-    @Test
-    public void execute_invalidIndexUnfilteredList_throwsCommandException() {
-        Index outOfBoundIndex = Index.fromOneBased(model.getFilteredPersonList().size() + 1);
-        InitDeleteCommand deleteCommand = new InitDeleteCommand(outOfBoundIndex);
-
-        assertCommandFailure(deleteCommand, model, Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
-    }
-
-    @Test
-    public void execute_invalidIndexFilteredList_throwsCommandException() {
-        showPersonAtIndex(model, INDEX_FIRST_PERSON);
-
-        Index outOfBoundIndex = INDEX_SECOND_PERSON;
-        // ensures that outOfBoundIndex is still in bounds of address book list
-        assertTrue(outOfBoundIndex.getZeroBased() < model.getAddressBook().getPersonList().size());
-
-        InitDeleteCommand deleteCommand = new InitDeleteCommand(outOfBoundIndex);
-
-        assertCommandFailure(deleteCommand, model, Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
-    }
-
-    @Test
-    public void equals() {
-        InitDeleteCommand initdeleteFirstCommand = new InitDeleteCommand(INDEX_FIRST_PERSON);
-        InitDeleteCommand initdeleteSecondCommand = new InitDeleteCommand(INDEX_SECOND_PERSON);
-
-        // same object -> returns true
-        assertTrue(initdeleteFirstCommand.equals(initdeleteFirstCommand));
-
-        // same values -> returns true
-        InitDeleteCommand deleteFirstCommandCopy = new InitDeleteCommand(INDEX_FIRST_PERSON);
-        assertTrue(initdeleteFirstCommand.equals(deleteFirstCommandCopy));
-
-        // different types -> returns false
-        assertFalse(initdeleteFirstCommand.equals(1));
-
-        // null -> returns false
-        assertFalse(initdeleteFirstCommand.equals(null));
-
-        // different person -> returns false
-        assertFalse(initdeleteFirstCommand.equals(initdeleteSecondCommand));
-    }
-
-    @Test
-    public void toStringMethod() {
-        Index targetIndex = Index.fromOneBased(1);
-        InitDeleteCommand deleteCommand = new InitDeleteCommand(targetIndex);
-        String expected = InitDeleteCommand.class.getCanonicalName() + "{targetIndex=" + targetIndex + "}";
-        assertEquals(expected, deleteCommand.toString());
-    }
-
-    /**
-     * Updates {@code model}'s filtered list to show no one.
-     */
-    private void showNoPerson(Model model) {
-        model.updateFilteredPersonList(p -> false);
-
-        assertTrue(model.getFilteredPersonList().isEmpty());
-    }
-}
