+++ src/test/java/seedu/hirehub/logic/parser/DeletePersonCommandParserTest.java
@@ -7,27 +7,26 @@ import static seedu.hirehub.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.hirehub.logic.commands.DeleteCommand;
-import seedu.hirehub.logic.commands.InitDeleteCommand;
+import seedu.hirehub.logic.commands.DeletePersonCommand;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
- * outside of the DeleteCommand code. For example, inputs "1" and "1 abc" take the
- * same path through the DeleteCommand, and therefore we test only one of them.
+ * outside of the DeletePersonCommand code. For example, inputs "1" and "1 abc" take the
+ * same path through the DeletePersonCommand, and therefore we test only one of them.
  * The path variation for those two cases occur inside the ParserUtil, and
  * therefore should be covered by the ParserUtilTest.
  */
-public class InitDeleteCommandParserTest {
+public class DeletePersonCommandParserTest {
 
-    private InitDeleteCommandParser parser = new InitDeleteCommandParser();
+    private final DeletePersonCommandParser parser = new DeletePersonCommandParser();
 
     @Test
     public void parse_validArgs_returnsDeleteCommand() {
-        assertParseSuccess(parser, "1", new InitDeleteCommand(INDEX_FIRST_PERSON));
+        assertParseSuccess(parser, "1", new DeletePersonCommand(INDEX_FIRST_PERSON));
     }
 
     @Test
     public void parse_invalidArgs_throwsParseException() {
-        assertParseFailure(parser, "a", String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+        assertParseFailure(parser, "a", String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeletePersonCommand.MESSAGE_USAGE));
     }
 }
